bot/handlers/dispatcher.py
"""
    Telegram event handlers
"""

import telegram
from telegram.ext import (
    Updater, Dispatcher, 
    CommandHandler,
    CallbackQueryHandler,
)
from .commands import command_start

from celery import task  # event processing in async mode

from conf.settings import TELEGRAM_TOKEN

from bot.handlers.button_dispatcher import button_dispatcher



def setup_dispatcher(dp):
    """
    Adding handlers for events from Telegram
    """

    dp.add_handler(CommandHandler("start", command_start))
    dp.add_handler(CallbackQueryHandler(button_dispatcher))

    return dp


def run_pooling():
    """ Run bot in pooling mode """
    updater = Updater(TELEGRAM_TOKEN, use_context=True)

    dp = updater.dispatcher
    dp = setup_dispatcher(dp)

    bot_info = telegram.Bot(TELEGRAM_TOKEN).get_me()
    bot_link = f"https://t.me/" + bot_info["username"]

    print(f"Pooling of '{bot_link}' started")
    updater.start_polling()
    updater.idle()


@task(ignore_result=True)
def process_telegram_event(update_json):
    update = telegram.Update.de_json(update_json, bot)
    dispatcher.process_update(update)


# Global variable - best way I found to init Telegram bot
bot = telegram.Bot(TELEGRAM_TOKEN)
dispatcher = setup_dispatcher(Dispatcher(bot, None, workers=0, use_context=True))
TELEGRAM_BOT_USERNAME = bot.get_me()["username"]