pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "boardroom-tg-bot"
version = "0.1.0"
authors = ["seyedrezafar <seyedrezafar@users.noreply.github.com>", "hhio618 <hhio618@users.noreply.github.com>"]
description = "Telegram bot for BroadRoom API"
repository = "https://github.com/seyedrezafar/boardroom-telegram-bot"
readme = "README.md"
license = "MIT"
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Topic :: Utilities",
    "Typing :: Typed",
]
keywords = ["bot", "boardroom", "API"]

packages = [
    {include = "bot"},
]
include = ["bot/py.typed"]

[tool.poetry.urls]
"Changelog" = "https://github.com/seyedrezafar/boardroom-telegram-bot/blob/master/CHANGELOG.md"

[tool.poetry.dependencies]
python = ">=3.6.2,<4"
python-telegram-bot = "*"
aiogram = "2.14.3"
aiohttp = "3.7.4.post0"
asgiref = "3.4.1"
async-timeout = "3.0.1"
attrs = "21.2.0"
Babel = "2.9.1"
certifi = "2021.5.30"
chardet = "4.0.0"
charset-normalizer = "2.0.4"
Django = "3.2.6"
idna = "3.2"
multidict = "5.1.0"
pytz = "2021.1"
requests = "2.26.0"
sqlparse = "0.4.1"
typing-extensions = "3.10.0.0"
urllib3 = "1.26.6"
yarl = "1.6.3"
celery = "4.3"
redis = "3.2.0"
django-celery-beat = "1.1.1"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
# Tests
pytest = "*"
coverage = { version = "*", extras = ["toml"] }

# Tools
mypy = "*"
pre-commit = "*"
bump2version = "*"
python-dotenv = "*"
ipdb = "*"
importlib-metadata = "*"
isort = "==4.3.19"
black = "==21.7b0"

[tool.poetry.build]
generate-setup-file = false

[tool.poe.tasks]
typecheck = "mypy ."
codestyle = "black --check --diff ."
check = ["typecheck", "codestyle"]

[tool.pytest.ini_options]
minversion = "2.0"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
python_classes = ["Test"]
python_functions = ["test"]
testpaths = ["tests"]
xfail_strict = true

[tool.black]
line-length = 99
multi_line_output=3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
exclude = '''
/(
  | codec
)/

'''


[tool.isort]
# Force imports to be sorted by module, independent of import type
force_sort_within_sections = true
# Group first party and local folder imports together
no_lines_before = ["LOCALFOLDER"]

# Configure isort to work without access to site-packages
known_first_party = ["bot", "tests"]

# Settings for Black compatibility
profile = "black"
line_length = 99
skip_glob = "^codec"



[tool.ward]
order = "random"


[tool.coverage.run]
omit = ["tests/*", "*/site-packages/*"]


[tool.mypy]
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
# Disabling incremental mode is required for `warn_unused_configs = true` to work
incremental = false
disallow_untyped_defs = true
check_untyped_defs = true
strict_equality = true
implicit_reexport = false
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "*.codec.*"
ignore_missing_imports = true
